Pull Make EML button out into separate component
